requirements.txt
annotated-types==0.7.0
anyio==4.10.0
asgiref==3.9.1
beautifulsoup4==4.13.5
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.2.1
colorama==0.4.6
cryptography==46.0.1
ddgs==9.5.4
deprecation==2.1.0
distlib==0.3.9
Django==5.2.6
django-cors-headers==4.9.0
djangorestframework==3.16.1
djangorestframework_simplejwt==5.5.1
dotenv==0.9.9
duckduckgo_search==8.1.1
filelock==3.18.0
fsspec==2025.7.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.34.4
hyperframe==6.1.0
idna==3.10
Jinja2==3.1.6
lxml==6.0.1
markdown-it-py==4.0.0
markdownify==1.2.0
MarkupSafe==3.0.2
mdurl==0.1.2
mysql-connector==2.2.9
ollama==0.5.3
packaging==25.0
pillow==11.3.0
platformdirs==4.3.8
postgrest==2.20.0
primp==0.15.0
psycopg==3.2.10
psycopg-binary==3.2.10
pycparser==2.23
pydantic==2.11.7
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
python-dotenv==1.1.1
PyYAML==6.0.2
realtime==2.20.0
requests==2.32.5
rich==14.1.0
six==1.17.0
smolagents==1.21.2
sniffio==1.3.1
soupsieve==2.8
sqlparse==0.5.3
storage3==2.20.0
StrEnum==0.4.15
supabase==2.20.0
supabase-auth==2.20.0
supabase-functions==2.20.0
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==2.5.0
virtualenv==20.31.2
websockets==15.0.1
google-generativeai==0.7.2
